write tests for the formatters
