--- test/Stubs.ts
@@ -1,3 +1,0 @@
-export interface FsStub {
-    writeFile: any;
-};
